Cleaning

+++ abcel/src/main/java/br/com/project_abcel/abcel/access/dto/FarmerDTO.java
@@ -4,6 +4,6 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
 public record FarmerDTO(@NotNull String name, String email, @NotNull String city, String neighborhood,
-		@NotBlank int phone) {
+		@NotBlank long phone) {
 
 }
